+++ solr/core/src/java/org/apache/solr/spelling/SpellCheckCollation.java
@@ -20,7 +20,7 @@ import org.apache.solr.common.util.NamedList;
 
 public class SpellCheckCollation implements Comparable<SpellCheckCollation> {
   private NamedList<String> misspellingsAndCorrections;
-  private int hits;
+  private long hits;
   private int internalRank;
   private String collationQuery;
 
@@ -42,11 +42,11 @@ public class SpellCheckCollation implements Comparable<SpellCheckCollation> {
     this.misspellingsAndCorrections = misspellingsAndCorrections;
   }
 
-  public int getHits() {
+  public long getHits() {
     return hits;
   }
 
-  public void setHits(int hits) {
+  public void setHits(long hits) {
     this.hits = hits;
   }
